requirements.txt
cromlech.marshallers == 0.1
git+https://github.com/HorsemanWSGI/roughrider.session.git@master#egg=roughrider.session
